pk: Assume UTF-8 encoding for JSON downloads
RFC 4627 states that "JSON text SHALL be encoded in Unicode.  The
default encoding is UTF-8."

Assume the use of UTF-8 encoding for JSON downloads, unless overridden
by the HTTP response headers.  This avoids the use of chardet to guess
the character set, which can be prohibitively expensive on large
files.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>